Fill in error-handling gaps
Errors from ftell/fseek not being handled in several places

+++ librxvm/bmh.h
@@ -13,6 +13,6 @@
 
 void bmh_reset (void);
 void bmh_init (FILE *file, char *pat, unsigned int length);
-int bmh (threads_t *tm);
+int64_t bmh (threads_t *tm);
 
 #endif
